project/config.py
import os 
from dotenv import load_dotenv
load_dotenv ()

DIR = "data"
Chroma_DB = "chroma_db"

EMBEDDING_MODEL = "sentence-transformers/all-MiniLM-L6-v2"
LLM_MODEL = "llama-3.1-8b-instant"

GROQ_API_KEY = os.getenv("GROQ_API_KEY")

if not GROQ_API_KEY:
    raise EnvironmentError("Groq api key is not found")